refactor: split Usage into usage.h so utils.h no longer depends on LibTorch.
