+++ xllm/core/framework/request/request_output.h
@@ -25,21 +25,9 @@ limitations under the License.
 
 #include "core/common/types.h"
 #include "core/framework/multimodal/embedding_output.h"
+#include "core/framework/request/usage.h"
 
 namespace xllm {
-struct Usage {
-  // the number of tokens in the prompt.
-  int32_t num_prompt_tokens = 0;
-
-  // the number of tokens in the generated completion.
-  int32_t num_generated_tokens = 0;
-
-  // the total number of tokens used in the request (prompt + completion).
-  int32_t num_total_tokens = 0;
-
-  // the number of prompt tokens served from prefix cache.
-  int32_t num_cached_tokens = 0;
-};
 
 struct LogProbData {
   // the text of the token.
